requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.0.0
Markdown==3.7
mdv==1.7.5
packaging==24.2
pluggy==1.5.0
Pygments==2.19.1
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
